requirements.txt
goatools==1.5.2
matplotlib==3.9.4
numpy==2.0.2
pandas==2.3.3
pronto==2.7.3
scikit_learn==1.6.1
tqdm==4.67.1
transformers==4.57.6
torch_geometric==2.6.1
torch_scatter==2.1.2